+++ week2/assignment/가계부 💸/js/utils/constants.js
